src/devices/arduino/arduino.py
"""
Arduino device controller.

This module provides the Arduino class for communicating with Arduino
microcontrollers via serial communication protocols.
"""

from typing import Any, Dict, Optional
import serial
import time
import logging


class Arduino:
    """
    Arduino microcontroller communication class.
    
    This class handles serial communication with Arduino devices,
    providing methods for sending commands and reading responses.
    
    Example:
        arduino = Arduino("lab_arduino_01", port="COM3", baudrate=9600)
        arduino.connect()
        value = arduino.read_analog_pin(0)
        arduino.disconnect()
    """
    
    def __init__(self, device_id: str, port: str, baudrate: int = 9600, 
                 timeout: float = 1.0, **kwargs):
        """
        Initialize Arduino device.
        
        Args:
            device_id: Unique identifier for the Arduino
            port: Serial port (e.g., 'COM3' on Windows, '/dev/ttyUSB0' on Linux)
            baudrate: Communication speed (default: 9600)
            timeout: Serial communication timeout in seconds
            **kwargs: Additional connection parameters
        """
        self.device_id = device_id
        self.port = port
        self.baudrate = baudrate
        self.timeout = timeout
        self.is_connected = False
        self.serial_connection: Optional[serial.Serial] = None
        self.logger = logging.getLogger(f"Arduino_{device_id}")
        
    def connect(self) -> bool:
        """
        Establish serial connection to the Arduino.
        
        Returns:
            bool: True if connection successful, False otherwise
        """
        try:
            # Implementation will be added here
            self.logger.info(f"Connecting to Arduino {self.device_id} on {self.port}")
            # self.serial_connection = serial.Serial(self.port, self.baudrate, timeout=self.timeout)
            self.is_connected = True
            return True
        except Exception as e:
            self.logger.error(f"Failed to connect to Arduino: {e}")
            return False
    
    def disconnect(self) -> bool:
        """
        Close serial connection to the Arduino.
        
        Returns:
            bool: True if disconnection successful, False otherwise
        """
        try:
            if self.serial_connection:
                self.serial_connection.close()
            self.is_connected = False
            self.logger.info(f"Disconnected from Arduino {self.device_id}")
            return True
        except Exception as e:
            self.logger.error(f"Failed to disconnect from Arduino: {e}")
            return False
    
    def get_status(self) -> Dict[str, Any]:
        """
        Get current Arduino status.
        
        Returns:
            Dict[str, Any]: Dictionary containing Arduino status information
        """
        return {
            "device_id": self.device_id,
            "port": self.port,
            "baudrate": self.baudrate,
            "connected": self.is_connected,
            "timeout": self.timeout
        }
    
    def send_command(self, command: str, wait_for_response: bool = True, 
                    response_timeout: float = None) -> Any:
        """
        Send a command to the Arduino.
        
        Args:
            command: Command string to send
            wait_for_response: Whether to wait for a response
            response_timeout: Timeout for response (uses default if None)
            
        Returns:
            Any: Response from the Arduino or None if no response expected
        """
        if not self.is_connected:
            self.logger.error("Arduino not connected")
            return None
            
        # Implementation will be added here
        self.logger.debug(f"Sending command: {command}")
        return f"Response to: {command}"  # Placeholder
    
    def read_analog_pin(self, pin: int) -> int:
        """
        Read analog value from specified pin.
        
        Args:
            pin: Analog pin number to read
            
        Returns:
            int: Analog value (0-1023 for 10-bit ADC)
        """
        # Implementation will be added here
        return 0
    
    def read_digital_pin(self, pin: int) -> bool:
        """
        Read digital value from specified pin.
        
        Args:
            pin: Digital pin number to read
            
        Returns:
            bool: Pin state (True for HIGH, False for LOW)
        """
        # Implementation will be added here
        return False
    
    def write_digital_pin(self, pin: int, value: bool) -> bool:
        """
        Write digital value to specified pin.
        
        Args:
            pin: Digital pin number to write
            value: Value to write (True for HIGH, False for LOW)
            
        Returns:
            bool: True if successful, False otherwise
        """
        # Implementation will be added here
        return True
    
    def write_pwm_pin(self, pin: int, value: int) -> bool:
        """
        Write PWM value to specified pin.
        
        Args:
            pin: PWM-capable pin number
            value: PWM value (0-255)
            
        Returns:
            bool: True if successful, False otherwise
        """
        # Implementation will be added here
        return True
