pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "esibd-lab-devices"
version = "0.1.0"
description = "Device management system for ESIBD laboratory equipment"
authors = [{name = "Niclas Przybylla", email = "your.email@example.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",        # Unit testing framework
    "pytest-cov>=4.0.0",    # Code coverage reports
    "black>=22.0.0",        # Code formatter
    "flake8>=5.0.0",        # Code linter
]
docs = [
    "sphinx>=5.0.0",             # Documentation generator
    "sphinx-rtd-theme>=1.0.0",   # Documentation theme
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
